+++ app/src/main/java/remix/myplayer/ui/appwidgets/medium/AppWidgetMedium.kt
@@ -51,7 +51,7 @@ class AppWidgetMedium : BaseAppwidget() {
     skin = AppWidgetSkin.WHITE_1F
     updateRemoteViews(service, remoteViews, song)
     //设置时间
-    val currentTime = progressState.position.toLong()
+    val currentTime = progressState.position
     val remainTime = song.duration - progressState.position
     if (currentTime > 0 && remainTime > 0) {
       remoteViews.setTextViewText(
@@ -76,7 +76,7 @@ class AppWidgetMedium : BaseAppwidget() {
     skin = AppWidgetSkin.WHITE_1F
     updateRemoteViews(service, remoteViews, song)
     //设置时间
-    val currentTime = progressState.position.toLong()
+    val currentTime = progressState.position
     val remainTime = song.duration - progressState.position
     if (currentTime > 0 && remainTime > 0) {
       remoteViews.setTextViewText(
